backend/requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
uvicorn==0.30.1
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
groq==0.9.0
fastembed==0.3.6
numpy==1.26.4
